pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "motion-planner"
version = "0.0.1"
requires-python = ">=3.8"
dependencies = ["numpy", "matplotlib", "python-fcl", "scipy"]